// packages/adapters/adapter-chargebee/src/generated/resource-specs.ts
export const chargebeeResourceSpecs: AdapterResourceSpecs = {
  platform: {
    timestampFormat: 'unix_seconds',
    amountFormat: 'integer_cents',
    idPrefix: '',
  },
  resources: {
// ...
    "coupon": {
      objectType: "coupon",
      volumeHint: "reference",
      refs: [],
      fields: {
        "id": { type: "string", required: true, default: "", idPrefix: "cb_cpn_" },
        "name": { type: "string", required: true, default: "" },
        "invoice_name": { type: "string", required: false, default: null },
        "discount_type": { type: "string", required: true, default: "fixed_amount", enum: ["fixed_amount","percentage","offer_quantity"] },
        "discount_percentage": { type: "number", required: false, default: 0 },
        "discount_amount": { type: "integer", required: false, default: 0 },
        "discount_quantity": { type: "integer", required: false, default: 0 },
        "currency_code": { type: "string", required: false, default: null, semanticType: "currency_code" },
        "duration_type": { type: "string", required: true, default: "one_time", enum: ["one_time","forever","limited_period"] },
        "valid_from": { type: "integer", required: false, default: 0 },
        "valid_till": { type: "integer", required: false, default: 0 },
        "max_redemptions": { type: "integer", required: false, default: 0 },
        "status": { type: "string", required: false, default: "active", enum: ["active","expired","archived","deleted","future"] },
        "apply_on": { type: "string", required: true, default: "invoice_amount", enum: ["invoice_amount","each_specified_item"] },
        "created_at": { type: "integer", required: true, auto: true, timestamp: "unix_seconds" },
        "archived_at": { type: "integer", required: false, auto: true, timestamp: "unix_seconds" },
        "resource_version": { type: "integer", required: false, auto: true, timestamp: "unix_seconds" },
        "updated_at": { type: "integer", required: false, auto: true, timestamp: "unix_seconds" },
        "period": { type: "integer", required: false, default: 0 },
        "period_unit": { type: "string", required: false, default: "day", enum: ["day","week","month","year"] },
        "redemptions": { type: "integer", required: false, default: 0 },
        "invoice_notes": { type: "string", required: false, default: null },
        "meta_data": { type: "object", required: false, default: {} },
        "deleted": { type: "boolean", required: true, default: false },
        "item_constraints": { type: "array", required: false, default: [] },
        "item_constraint_criteria": { type: "array", required: false, default: [] },
        "coupon_constraints": { type: "array", required: false, default: [] },
      },
    },
// ...
  },
};